+++ lib/rodbot/db/redis.rb
@@ -11,7 +11,7 @@ module Rodbot
     # @example Enable in config/rodbot.rb
     #   db 'redis://localhost:6379/10'
     module Redis
-      include Rodbot::Concerns::Memoize
+      include Rodbot::Memoize
 
       def self.extended(*)
         require 'redis'