+++ spec/lib/rodbot/services/relay_spec.rb
@@ -7,7 +7,7 @@ describe Rodbot::Services::Relay do
 
   describe :url do
     it "returns tcp://localhost and ports above 7200 by default" do
-      Rodbot::Concerns::Memoize::suspend do
+      Rodbot::Memoize::suspend do
         with '@config', Rodbot::Config.new("plugin :matrix; plugin :slack"), on: Rodbot do
           _(subject.url(:matrix)).must_equal 'tcp://localhost:7201'
           _(subject.url(:slack)).must_equal 'tcp://localhost:7202'
@@ -16,7 +16,7 @@ describe Rodbot::Services::Relay do
     end
 
     it "returns tcp://localhost and ports above explicit port config" do
-      Rodbot::Concerns::Memoize::suspend do
+      Rodbot::Memoize::suspend do
         with '@config', Rodbot::Config.new("plugin :matrix; plugin :slack; port 8888"), on: Rodbot do
           _(subject.url(:matrix)).must_equal 'tcp://localhost:8889'
           _(subject.url(:slack)).must_equal 'tcp://localhost:8890'
@@ -25,7 +25,7 @@ describe Rodbot::Services::Relay do
     end
 
     it "returns value of RODBOT_RELAY_URL_XXX and ports above 7200" do
-      Rodbot::Concerns::Memoize::suspend do
+      Rodbot::Memoize::suspend do
         with "ENV['RODBOT_RELAY_URL_MATRIX']", 'tcp://matrix.relay.local' do
           with '@config', Rodbot::Config.new("plugin :matrix; plugin :slack"), on: Rodbot do
             _(subject.url(:matrix)).must_equal 'tcp://matrix.relay.local:7201'