+++ CHANGELOG.md
@@ -5,6 +5,6 @@
 * App, relay and schedule services
 * Framework functionality such as config, credentials, logging, data
   persistence, async jobs and so forth
-* Refinements and concerns
+* Refinements and memoization
 * Built-in plugins for Matrix, GitHub, GitLab, OTP and some demos
 * Require at least Ruby 3.0